images/auth/pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.1",
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.120.0",
    "orjson>=3.11.4",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "pymongo>=4.15.5",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.0.0",
    "sqlalchemy>=2.0.44",
    "user-agents>=2.2.0",
    "uvicorn[standard]>=0.38.0",
    "werkzeug>=3.1.3",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "flake8>=7.3.0",
    "flake8-pyproject>=1.2.3",
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "wemake-python-styleguide>=1.4.0",
]

[tool.flake8]
max-line-length = 100
exclude = [".venv", "*/migrations/*.py"]

[tool.isort]
line_length = 100
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
skip_glob = [".venv"]
default_section = "THIRDPARTY"
src_paths = ["src"]
known_flask = ["flask"]
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "FLASK",
    "FIRSTPARTY",
    "LOCALFOLDER"
]

[tool.black]
line-length = 100
exclude = '\.venv/'
skip-string-normalization = true

[tool.mypy]
python_version = "3.13"
strict = true
show_error_codes = true
warn_unused_ignores = true
exclude = "(^|/)(.venv|migrations)/"
namespace_packages = true
ignore_missing_imports = true

[tool.uv.sources]
auth = { workspace = true }
